server/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
cffi==1.16.0
cryptography==42.0.5
exceptiongroup==1.2.0
fastapi==0.109.2
idna==3.6
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
